backend/notification/notification/requirements.txt
amqp==5.3.1
asgiref==3.9.2
async-timeout==5.0.1
bidict==0.23.1
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
eventlet==0.40.3
greenlet==3.2.4
h11==0.16.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.14.0
redis==6.4.0
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14
wsproto==1.2.0
